+++ src/include/noisyheight_editor.h
@@ -12,7 +12,7 @@ class NoisyHeight_Editor : public Editor
 {
 public:
     NoisyHeight_Editor(Shape* shape);
-    NoisyHeight_Editor(Shape* shape, int octave, int frequence, int persistence);
+    NoisyHeight_Editor(Shape* shape, int octave, float frequence, float persistence);
     ~NoisyHeight_Editor();
 
     void edit();
@@ -23,8 +23,8 @@ private:
     void assignColor(Shape::Vertices* vertices, Eigen::Vector3f colors);
 
     int _octave;
-    int _frequence;
-    int _persistence;
+    float _frequence;
+    float _persistence;
 };
 
 #endif // NOISYHEIGHT_EDITOR_H
